+++ resources/views/layouts/app/sidebar.blade.php
@@ -15,19 +15,39 @@
                     <flux:sidebar.item icon="home" :href="route('dashboard')" :current="request()->routeIs('dashboard')" wire:navigate>
                         {{ __('Dashboard') }}
                     </flux:sidebar.item>
+
+                    <flux:sidebar.item icon="photo" href="#">
+                        {{ __('Content') }}
+                    </flux:sidebar.item>
+
+                    <flux:sidebar.item icon="calendar" href="#">
+                        {{ __('Schedule') }}
+                    </flux:sidebar.item>
+
+                    <flux:sidebar.item icon="chart-bar" href="#">
+                        {{ __('Analytics') }}
+                    </flux:sidebar.item>
                 </flux:sidebar.group>
-            </flux:sidebar.nav>
 
-            <flux:spacer />
+                <flux:sidebar.group :heading="__('Manage')" class="grid">
+                    <flux:sidebar.item icon="users" href="#">
+                        {{ __('Clients') }}
+                    </flux:sidebar.item>
 
-            <flux:sidebar.nav>
-                <flux:sidebar.item icon="folder-git-2" href="https://github.com/laravel/livewire-starter-kit" target="_blank">
-                    {{ __('Repository') }}
-                </flux:sidebar.item>
+                    <flux:sidebar.item icon="document-text" href="#">
+                        {{ __('Proposals') }}
+                    </flux:sidebar.item>
 
-                <flux:sidebar.item icon="book-open-text" href="https://laravel.com/docs/starter-kits#livewire" target="_blank">
-                    {{ __('Documentation') }}
-                </flux:sidebar.item>
+                    <flux:sidebar.item icon="banknotes" href="#">
+                        {{ __('Invoices') }}
+                    </flux:sidebar.item>
+                </flux:sidebar.group>
+
+                <flux:sidebar.group :heading="__('Instagram')" class="grid">
+                    <flux:sidebar.item icon="at-symbol" href="#">
+                        {{ __('Accounts') }}
+                    </flux:sidebar.item>
+                </flux:sidebar.group>
             </flux:sidebar.nav>
 
             <x-desktop-user-menu class="hidden lg:block" :name="auth()->user()->name" />